update prep_build.bash; handle NULL options in tidypath

+++ prep_build.bash
@@ -5,4 +5,14 @@
 # as `prep_build.bash --force` to force a rebuild of everything.
 #
 
+set -o errexit
+
+if test ! -f README.md; then
+    echo "prep_build.bash must be run from the top level directory" 1>&2
+    exit 1
+fi
+
+# reduce warnings by pre-creating the m4 directory
+mkdir -p m4
+
 exec autoreconf --install "$@"
